mfd: lan966x: add pci driver
Add a PCI driver that handles the lan966x boards using a device-tree
overlay. This overlay is applied to the PCI device node and allows
describing the whole SoC that is present on the card. The memory from
the device-tree are remapped to the BAR memory thanks to added "ranges"
properties.

Signed-off-by: Clément Léger <clement.leger@bootlin.com>