net: stmmac: start phylink before setting up hardware
The stmmac on the Renesas RZ/N1 platform is connected to the PCS which must be
configured to provide a correct RGMII RX clock to the stmmac IP. Without the RX
clock, the driver will fail to initialize the hardware (More specifically, the
driver will report it fails to reset DMA). In order to fix that, start phylink
mecanism before setting up hardware.

Signed-off-by: Clément Léger <clement.leger@bootlin.com>