pci: add pci_get_of_node()
Add pci_get_of_node() which allows to get or create the node and parent
nodes associated to the PCI device. This function is meant to be used
by PCI drivers that want to apply a device-tree overlay on a PCI device.
If the device-tree node does not exists, this function will take care
of creating a correct tree allowing to insert the overlay.

Signed-off-by: Clément Léger <clement.leger@bootlin.com>